+++ source/core/tests/unit_tests/gold/infrastructure/streams/test_measurements_gold_stream.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 from pytest_mock import MockFixture
 
-from core.gold.infrastructure.repositories.gold_measurements_repository import GoldMeasurementsRepository
+from core.gold.infrastructure.streams.gold_measurements_stream import GoldMeasurementsStream
 
 
 def test__start_write_stream__calls_expected(
@@ -14,7 +14,7 @@ def test__start_write_stream__calls_expected(
     mocked_batch_operation = mock.Mock()
     mocker.patch("core.gold.infrastructure.config.spark.initialize_spark")
     # Act
-    GoldMeasurementsRepository().write_stream(
+    GoldMeasurementsStream().write_stream(
         mocked_measurements,
         mocked_batch_operation,
     )
@@ -40,7 +40,7 @@ def test__start_write_stream__when_contionous_streaming_is_disabled__should_not_
     mocked_batch_operation = mock.Mock()
     mocker.patch("core.gold.infrastructure.config.spark.initialize_spark")
     # Act
-    GoldMeasurementsRepository().write_stream(mocked_measurements, mocked_batch_operation)
+    GoldMeasurementsStream().write_stream(mocked_measurements, mocked_batch_operation)
 
     # Assert
     mocked_measurements.writeStream.format().queryName().option().trigger.assert_not_called()
